upgraded ncbi-vdb to 2.5.7
the opam-install script was also modified to ease the installation of
upcoming package ncbi-sra-tools. Here's a comment related to the
change:

The build of package ncbi-sra-tools requires source and build results
of ncbi-vdb. This is a vestige from the time all three ncbi libraries
(ngs, vdb and sra-tools) were a single entity. It seems it should be
fixed in the long run:

  https://github.com/ncbi/ncbi-vdb/issues/3