ncbi-{vdb,sra-tools}: add libhdf5 as depext
ideally libhdf5 should only be an optional depext for both packages,
since they can be built without it. The thing is vdb produces an
additional library and sra-tools produces additional executables so we
would have to take care of this for the ".install" file. Since there
will probably be major upstream changes on these packages, I think we
can see that another time.